pyproject.toml
[project]
name = "track-money"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.136.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.11",
]
